+++ src/AlmatyLISPollingBot.Application/Contracts/Bot/BotCommands.cs
@@ -6,6 +6,7 @@ public static class BotCommands
     public const string Options = "options";
     public const string Exclude = "exclude";
     public const string Excluded = "excluded";
+    public const string Unexclude = "unexclude";
     public const string Force = "force";
     public const string Cancel = "cancel";
     public const string Stop = "stop";