+++ src/AlmatyLISPollingBot.Worker/Telegram/PrivateAdminDialogKind.cs
@@ -3,5 +3,6 @@ namespace AlmatyLISPollingBot.Worker.Telegram;
 public enum PrivateAdminDialogKind
 {
     ExcludeTournaments,
+    UnexcludeTournaments,
     ForceTournaments
 }